Extract evm TS logic to evm-service
