+++ packages/evm-service/README.md
@@ -1,0 +1,19 @@
+# Mainsail - Transaction pool worker
+
+![banner](https://raw.githubusercontent.com/ArkEcosystem/mainsail/main/banner.jpeg)
+
+## Documentation
+
+You can find installation instructions and detailed instructions on how to use this package at the [dedicated documentation site](https://ark.dev/docs/mainsail).
+
+## Security
+
+If you discover a security vulnerability within this package, please send an e-mail to [security@ark.io](mailto:security@ark.io). All security vulnerabilities will be promptly addressed.
+
+## Credits
+
+This project exists thanks to all the people who [contribute](https://github.com/ArkEcosystem/mainsail/graphs/contributors).
+
+## License
+
+[GPL-3.0-only](https://github.com/ArkEcosystem/mainsail/blob/main/LICENSE) © [ARK Ecosystem](https://ark.io)
