+++ packages/evm-service/tsconfig.json
@@ -3,5 +3,5 @@
 	"compilerOptions": {
 		"outDir": "distribution"
 	},
-	"include": ["source/**/**.ts", "source/**/**.d.cts"]
+	"include": ["source/**/**.ts"]
 }
