+++ packages/evm/package.json
@@ -4,15 +4,12 @@
 	"description": "EVM for the Mainsail blockchain",
 	"license": "GPL-3.0-only",
 	"contributors": [],
-	"type": "module",
-	"main": "distribution/index.js",
-	"types": "distribution/index.d.ts",
-	"files": [
-		"distribution"
-	],
+	"main": "index.js",
+	"types": "index.d.ts",
 	"scripts": {
-		"build": "pnpm run clean && pnpm run build-napi && tsc && pnpm run copy",
+		"build": "pnpm run clean && pnpm run build-napi2",
 		"build-napi": "pnpm run clean && mkdir source/generated && napi build --platform --cargo-cwd bindings --js bindings.cjs --dts bindings.d.cts --release && mv evm.*.node bindings.cjs bindings.d.cts source/generated",
+		"build-napi2": "napi build --platform --cargo-cwd bindings --release",
 		"build-napi:debug": "napi build --platform --cargo-cwd bindings --js bindings.cjs --dts bindings.d.cts && mv evm.*.node bindings.cjs bindings.d.cts source/generated",
 		"build:watch": "pnpm run clean && tsc -w",
 		"clean": "del distribution && del source/generated",
